Adding basic resource implementations, for Group and Object.
'describe' methods on Groups and Objects are reified as GETs.
'register*' operations are POSTs.

Added a dependency on Google Guice to handle dependency injection on the resource objects themselves.
The (internal) BOSS API is written as an interface, and a simple (in-memory) implementation is provided.

Basic (HTTP) tests are provided for the register and describe functions.