[DEV-1330] Database support for persistent resources across restarts.
This commit includes support for using a database to persist [Fs]Object/Group resources, so that the state of the API
can be preserved across restarts (or between multiple API instances).  Also included is support for a behind-the-scenes
'location' field, attached to both Groups and Objects, which can store the location in an object store or filesystem
for a particular resource.  This is hooked up and filled in, but there aren't any actual methods for manipulating or
retrieving it -- the only way currently spec'ed out is through the presigned URL method in the future.

Basic outline for the changes here:
- added dependencies on HSQLDB (for testing), and using JDBI and Liquibase from the Dropwizard packages for the JDBC
  and database migration work.
- Added a BossDAO object which contains the SQL queries and mediates the interaction with the database.
- Added a new implementation of the API interface, DatabaseBossAPI, which wraps the DAO and provides access
  to the database for the resource classes.

The Fs* (filesystem) resources are supported through mapping their fields into the same groups/objects tables in the
database -- this is because I anticipate that supporting these resources as separate types/classes in the future will
probably cease to be a requirement.

Removed some of the outdated fields (*Url) from Object and Resource, and completed the transition to using exclusively
relative names in the groupId/objectId fields of all the resources.

Also included unit tests, and a script to manage the migration-prior-to-deploy necessary for running this in TEST or PROD.

Updated the .yml configuration file to support database configuration.