[DEV-1335] Anonymous Group/Object registration.
This commit satisfies DEV-1335, the story about creating Objects and Groups
(and FsObjects and FsGroups) without having an object ID or a group ID already
in-hand.

We've added four new endpoints, the first two for "anonymous group" creation:

/groups/store
/groups/fs

which take POSTs of [Fs]GroupResource values, generate random IDs, and register
the groups under those IDs.

The second two endpoints are for anonymous object creation,

/group/store/{groupId}/objects
/group/fs/{groupId}/objects

which take POSTs of [Fs]ObjectResource values, generate random IDs, and register
the object under that ID with the corresponding Group.

Finally, the URLs of _existing_ groups have been moved to

/group/store/{groupId}/object/{objectId}
/group/fs/{groupId}/object/{objectId}

so that they won't conflict with the new {groupId}/objects path.

Tests have been added for each of these calls, to AllGroupsAcceptanceTest
and the existing [Fs]GroupResourceAcceptanceTest classes.