[DEV-1329] Supporting object store configuration and URL generation
The Boss API is now configurable to run on top of a particular object store (specified
by an endpoint) and a particular bucket within that object store.  It generates random
locations within the bucket, and creates presigned URLs with respect to those locations.

This commit therefore also fixes DEV-1331.

We currently only support S3-compatible object stores, but there's enough abstracting
happening in here that we could support other URL-generating interfaces in the future.

Required the following additional changes:

* Adding dependency on AWS S3 SDK, but removing the conflicting Jackson libraries
  from the imported library.
* Changing the signature of the BossAPI method for generating pre-signed URLs.
* Defining a new configuration section within the configuration Yaml file, for
  managing the configuration (and bucket) of the object store on which we're running.
* Updated the 'resolve' method on the ObjectResource to take a 'seconds' parameter,
  which then dictates how long the URL should remain active for.