Fixing inclusions to remove warning when building uber-jar.
I'm leaving this as a separate commit, so we can see the changes explicitly.  This
fixes two warnings we were getting from the Shade maven plugin that was building the uber-jar.
See the notes in the parent pom.xml, but in short: two of the dependencies (servlet-api and
jsr305/annotations) were having two slightly different versions pulled in by two different
dropwizard components.

This commit does the right dependency/exclusion munging to fix those warnings.