[DEV-1393] Resolve will succeed on 'filesystem' objects.
Old code was still in the ObjectResource, dictating that only 'objectstore' resources would have a working resolve method.

This has now been fixed -- when a non-objectstore object is resolved, the directoryPath is turned into a file://-schemed URL.

I've added a test to make sure this particular behavior sticks around.